core: stub out ResourceProvider API for configschema
In the new configuration architecture core needs to know the schema of
a resource type or data source in order to decode its configuration. These
new ResourceProvider interface methods allow the provider to expose
this information.

This change is carefully defined to avoid the need to increase the plugin
protocol version, since the new method will only be called if the
provider declares support for it by returning SchemaAvailable as true
when reporting on its resources and data source.

Currently the helper/schema implementations of these new methods are
stubs. A later commit will add the implementation.