config: start of zcl-based RawConfig processing
The zcl usage model is very different than HCL/HIL since it requires
a schema to be defined before any processing can be done. Therefore the
existing RawConfig API is inappropriate for zcl.

However, RawConfig is prolific in Terraform's internal APIs right now, so
to allow a first pass experimental implementation of zcl parsing we'll
smuggle a zcl.Body inside an otherwise-degenerate RawConfig, pass that
down through the API, and then eventually recognize it at interpolation
time and branch into the new, separate processing codepath.

This is a temporary state. Eventually the API of the config package will
change to be zcl-first and the HCL/HIL facilities will be removed, along
with this hack.